+++ junos_ops/__init__.py
@@ -17,4 +17,4 @@ See Also:
     https://github.com/shigechika/junos-ops
 """
 
-__version__ = "0.10.1"
+__version__ = "0.11.0"